Replace handleServiceAgents
