+++ src/scenes/TransportationServiceProvider/ShipmentInfo.jsx
@@ -32,7 +32,11 @@ import {
   getPublicShipmentLabel,
   acceptShipment,
 } from 'shared/Entities/modules/shipments';
-import { getServiceAgentsForShipment, selectServiceAgentsForShipment } from 'shared/Entities/modules/serviceAgents';
+import {
+  getServiceAgentsForShipment,
+  selectServiceAgentsForShipment,
+  handleServiceAgents,
+} from 'shared/Entities/modules/serviceAgents';
 
 import FontAwesomeIcon from '@fortawesome/react-fontawesome';
 import faPhone from '@fortawesome/fontawesome-free-solid/faPhone';
@@ -45,7 +49,7 @@ import Dates from 'shared/ShipmentDates';
 import LocationsContainer from 'shared/LocationsPanel/LocationsContainer';
 import { getLastRequestIsSuccess, getLastRequestIsLoading, getLastError } from 'shared/Swagger/selectors';
 import { resetRequests } from 'shared/Swagger/request';
-import { completePmSurvey, transportShipment, deliverShipment, handleServiceAgents } from './ducks';
+import { completePmSurvey, transportShipment, deliverShipment } from './ducks';
 import FormButton from './FormButton';
 import CustomerInfo from './CustomerInfo';
 import PreApprovalPanel from 'shared/PreApprovalRequest/PreApprovalPanel.jsx';
