+++ src/scenes/TransportationServiceProvider/api.js
@@ -64,27 +64,6 @@ export async function CompletePmSurvey(shipmentId) {
 }
 
 // ServiceAgents
-export async function CreateServiceAgent(shipmentId, payload) {
-  const client = await getPublicClient();
-  const response = await client.apis.service_agents.createServiceAgent({
-    shipmentId,
-    serviceAgent: payload,
-  });
-  checkResponse(response, 'failed to create service agent due to server error');
-  return response.body;
-}
-
-export async function UpdateServiceAgent(payload) {
-  const client = await getPublicClient();
-  const response = await client.apis.service_agents.patchServiceAgent({
-    shipmentId: payload.shipment_id,
-    serviceAgentId: payload.id,
-    serviceAgent: payload,
-  });
-  checkResponse(response, 'failed to update service agent due to server error');
-  return response.body;
-}
-
 export async function IndexServiceAgents(shipmentId) {
   const client = await getPublicClient();
   const response = await client.apis.service_agents.indexServiceAgents({
