+++ src/shared/Entities/modules/serviceAgents.js
@@ -19,11 +19,40 @@ export function updateServiceAgentForShipment(
   return swaggerRequest(getPublicClient, swaggerTag, { shipmentId, serviceAgentId, serviceAgent }, { label });
 }
 
+export function createServiceAgentForShipment(shipmentId, serviceAgent, label = updateServiceAgentForShipmentLabel) {
+  const swaggerTag = 'service_agents.createServiceAgent';
+  return swaggerRequest(getPublicClient, swaggerTag, { shipmentId, serviceAgent }, { label });
+}
+
 export function updateServiceAgentsForShipment(shipmentId, serviceAgents, label = updateServiceAgentForShipmentLabel) {
   return async function(dispatch) {
     Object.values(serviceAgents).map(serviceAgent =>
       dispatch(updateServiceAgentForShipment(shipmentId, serviceAgent.id, serviceAgent, label)),
     );
+  };
+}
+
+export function handleServiceAgents(shipmentId, serviceAgents) {
+  return async function(dispatch) {
+    for (const serviceAgent in serviceAgents) {
+      /* eslint-disable security/detect-object-injection */
+      dispatch(createOrUpdateServiceAgent(shipmentId, serviceAgents[serviceAgent]));
+      /* eslint-enable security/detect-object-injection */
+    }
+  };
+}
+
+export function createOrUpdateServiceAgent(shipmentId, serviceAgent) {
+  return async function(dispatch, getState) {
+    if (serviceAgent.id) {
+      return dispatch(updateServiceAgentForShipment(shipmentId, serviceAgent.id, serviceAgent));
+    } else if (!serviceAgent.company || !serviceAgent.email || !serviceAgent.phone_number) {
+      // Don't send the service agent if it's not got enough details
+      // Currently, it should only be the destination agent that gets skipped
+      return;
+    } else {
+      return dispatch(createServiceAgentForShipment(shipmentId, serviceAgent));
+    }
   };
 }
